requirements.txt
gradio>=4.44
requests>=2.32
Pillow>=10.3
pandas>=2.2
python-dotenv>=1.0